refactor(settings): extract settings to layered architecture
- Create api/settings/admin.rs and api/settings/public.rs for handlers
- Create services/settings_service.rs for business logic
- Create repositories/settings_repository.rs for data access
- Remove settings functions from db/queries.rs
- Remove admin settings handlers from router.rs
- Apply AppError pattern to admin settings handlers
- Update lib.rs and openapi.rs to use new module paths
