--- src/api/settings.rs
@@ -1,44 +1,0 @@
-use crate::db;
-use worker::*;
-
-#[utoipa::path(
-    get,
-    path = "/api/settings/public",
-    tag = "Settings",
-    summary = "Get public settings",
-    description = "Returns non-sensitive public settings needed by the frontend, including founder pricing status and active discount amounts for each plan tier",
-    responses(
-        (status = 200, description = "Public settings object"),
-    )
-)]
-pub async fn handle_get_public_settings(
-    _req: Request,
-    ctx: worker::RouteContext<()>,
-) -> Result<Response> {
-    let db = ctx.env.get_binding::<worker::d1::D1Database>("rushomon")?;
-
-    // Get all settings in a single query (much faster than running individual queries)
-    let settings = db::get_all_settings(&db).await?;
-
-    // Get founder pricing status from settings
-    let founder_pricing_active = settings
-        .get("founder_pricing_active")
-        .map(|v| v == "true")
-        .unwrap_or(false);
-
-    // Helper to parse setting as i64
-    let get_setting_i64 = |key: &str| -> i64 {
-        settings
-            .get(key)
-            .and_then(|v| v.parse::<i64>().ok())
-            .unwrap_or(0)
-    };
-
-    Response::from_json(&serde_json::json!({
-        "founder_pricing_active": founder_pricing_active,
-        "active_discount_amount_pro_monthly": get_setting_i64("active_discount_amount_pro_monthly"),
-        "active_discount_amount_pro_annual": get_setting_i64("active_discount_amount_pro_annual"),
-        "active_discount_amount_business_monthly": get_setting_i64("active_discount_amount_business_monthly"),
-        "active_discount_amount_business_annual": get_setting_i64("active_discount_amount_business_annual"),
-    }))
-}
