+++ src/api/settings/mod.rs
@@ -1,0 +1,5 @@
+/// Settings API handlers
+///
+/// Public and admin endpoints for managing system settings.
+pub mod admin;
+pub mod public;
