+++ src/services/mod.rs
@@ -13,5 +13,7 @@
 // - Direct DB access (use repositories instead)
 
 // Add service modules here as they are created:
+pub mod settings_service;
 pub mod tag_service;
+pub use settings_service::SettingsService;
 pub use tag_service::TagService;
